Created medication linking logic
NOTE: During this process a bug was spotted where the viewable list of trees did not update
during the retrieval of pillpack data. This issue has now been rectified with the update_list_of_trees
function, which is called in the update function.

A medication can now be linked to another medication. This feature is to compensate for branded
medications being in production, but generics being on the script. This feature should NOT be used
the other way around (i.e. linking a generic to a branded drug on the script).
This link can also be severed by pressing the unlink button.