Fixed dosage inaccruacy for multiple entries
Pillpac care, for whatever reason, does not let you put through an entry with uneven
amounts (e.g. 2 in the morning, 1 in the afternoon, 2 at night) which means they have to be put through
as separate entries. I overlooked this when taking in production data. This commit should ensure that
if multiple entries under the same medication exist, the total dosage will be updated under the medication object
already stored in the patient's production medication dictionary.